organized related functions in same files
