+++ pointers/numeric.go
@@ -89,3 +89,39 @@ func NullableUint64(i *uint64) uint64 {
 	}
 	return *i
 }
+
+// NullableFloat32 returns the dereferenced value of *float32 if not nil,
+// or 0.0 otherwise.
+func NullableFloat32(f *float32) float32 {
+	if f == nil {
+		return 0.0
+	}
+	return *f
+}
+
+// NullableFloat64 returns the dereferenced value of *float64 if not nil,
+// or 0.0 otherwise.
+func NullableFloat64(f *float64) float64 {
+	if f == nil {
+		return 0.0
+	}
+	return *f
+}
+
+// NullableComplex64 returns the dereferenced value of *complex64 if not nil,
+// or 0+0i otherwise.
+func NullableComplex64(c *complex64) complex64 {
+	if c == nil {
+		return 0 + 0i
+	}
+	return *c
+}
+
+// NullableComplex128 returns the dereferenced value of *complex128 if not nil,
+// or 0+0i otherwise.
+func NullableComplex128(c *complex128) complex128 {
+	if c == nil {
+		return 0 + 0i
+	}
+	return *c
+}
