+++ pointers/numeric_test.go
@@ -331,3 +331,163 @@ func TestNullableUint64(t *testing.T) {
 		})
 	}
 }
+
+func TestNullableFloat32(t *testing.T) {
+	type args struct {
+		f *float32
+	}
+	tests := []struct {
+		name string
+		args args
+		want float32
+	}{
+		{
+			name: "success - f is nil",
+			args: args{
+				f: nil,
+			},
+			want: 0.0,
+		},
+		{
+			name: "success - f has a value",
+			args: args{
+				f: func() *float32 { v := float32(3.14); return &v }(),
+			},
+			want: 3.14,
+		},
+		{
+			name: "success - f is zero",
+			args: args{
+				f: func() *float32 { v := float32(0.0); return &v }(),
+			},
+			want: 0.0,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := NullableFloat32(tt.args.f); got != tt.want {
+				t.Errorf("NullableFloat32() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestNullableFloat64(t *testing.T) {
+	type args struct {
+		f *float64
+	}
+	tests := []struct {
+		name string
+		args args
+		want float64
+	}{
+		{
+			name: "success - f is nil",
+			args: args{
+				f: nil,
+			},
+			want: 0.0,
+		},
+		{
+			name: "success - f has a value",
+			args: args{
+				f: func() *float64 { v := float64(6.28); return &v }(),
+			},
+			want: 6.28,
+		},
+		{
+			name: "success - f is zero",
+			args: args{
+				f: func() *float64 { v := float64(0.0); return &v }(),
+			},
+			want: 0.0,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := NullableFloat64(tt.args.f); got != tt.want {
+				t.Errorf("NullableFloat64() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestNullableComplex64(t *testing.T) {
+	type args struct {
+		c *complex64
+	}
+	tests := []struct {
+		name string
+		args args
+		want complex64
+	}{
+		{
+			name: "success - c is nil",
+			args: args{
+				c: nil,
+			},
+			want: 0 + 0i,
+		},
+		{
+			name: "success - c is not nil and is 0+0i",
+			args: args{
+				c: func() *complex64 { v := complex64(0 + 0i); return &v }(),
+			},
+			want: 0 + 0i,
+		},
+		{
+			name: "success - c is not nil and is 1+2i",
+			args: args{
+				c: func() *complex64 { v := complex64(1 + 2i); return &v }(),
+			},
+			want: 1 + 2i,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := NullableComplex64(tt.args.c); got != tt.want {
+				t.Errorf("NullableComplex64() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestNullableComplex128(t *testing.T) {
+	type args struct {
+		c *complex128
+	}
+	tests := []struct {
+		name string
+		args args
+		want complex128
+	}{
+		{
+			name: "success - c is nil",
+			args: args{
+				c: nil,
+			},
+			want: 0 + 0i,
+		},
+		{
+			name: "success - c is not nil and is 0+0i",
+			args: args{
+				c: func() *complex128 { v := complex128(0 + 0i); return &v }(),
+			},
+			want: 0 + 0i,
+		},
+		{
+			name: "success - c is not nil and is 3+4i",
+			args: args{
+				c: func() *complex128 { v := complex128(3 + 4i); return &v }(),
+			},
+			want: 3 + 4i,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := NullableComplex128(tt.args.c); got != tt.want {
+				t.Errorf("NullableComplex128() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
